requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
schedule==1.2.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pyodbc==5.2.0
pymysql==1.1.0
